COM-2674 move mockCurrentDate
