+++ tests/unit/mockCurrentDate.js
@@ -1,0 +1,9 @@
+const luxon = require('../../src/helpers/dates/luxon');
+
+const mockCurrentDate = (expectedNowISO) => {
+  const luxonExpectedNow = luxon.DateTime.fromISO(expectedNowISO);
+
+  luxon.Settings.now = () => luxonExpectedNow.toMillis();
+};
+
+module.exports = { mockCurrentDate };
